WIP: save current working copy and subsystems before test fixes

+++ core/sh2/ops_arith.go
@@ -1,4 +1,4 @@
-// Copyright 2026 The erings Authors
+// Copyright 2026 The cassini Authors
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 package sh2
@@ -464,7 +464,7 @@ func opMACW(c *CPU) {
 	m := regM(c.ir)
 
 	// Cycle 1: read @Rn, post-increment
-	c.pendingVal = uint32(c.Read16(c.reg.R[n]))
+	c.pendingVal = uint32(int32(int16(c.Read16(c.reg.R[n]))))
 	c.reg.R[n] += 2
 	c.pendingN = n | (m << 4)
 	c.stepBus = BusRead
